moved shared, deps and features. added middlewares folder
