+++ express/components/carousel.js
@@ -1,4 +1,4 @@
-import { createTag, loadStyle } from '../../scripts/utils.js';
+import { createTag, loadStyle } from '../utils.js';
 
 function correctCenterAlignment(plat) {
   if (plat.parentElement.offsetWidth <= plat.offsetWidth) return;
